+++ Sources/VComponents/Components/State Pickers/Check Box/VCheckBox.swift
@@ -31,8 +31,8 @@ import VCore
 ///         )
 ///     }
 ///
-@available(tvOS, unavailable) // No `SwiftUIGestureBaseButton`
-@available(watchOS, unavailable) // No `SwiftUIGestureBaseButton`
+@available(tvOS, unavailable) // Doesn't follow HIG
+@available(watchOS, unavailable) // Doesn't follow HIG
 @available(visionOS, unavailable) // Doesn't follow HIG
 public struct VCheckBox<Label>: View where Label: View {
     // MARK: Properties - UI Model
@@ -40,13 +40,12 @@ public struct VCheckBox<Label>: View where Label: View {
 
     // MARK: Properties - State
     @Environment(\.isEnabled) private var isEnabled: Bool
-    @State private var isPressed: Bool = false
     @Binding private var state: VCheckBoxState
-    private var internalState: VCheckBoxInternalState {
+    private func internalState(_ baseButtonState: SwiftUIBaseButtonState) -> VCheckBoxInternalState {
         .init(
             isEnabled: isEnabled,
             state: state,
-            isPressed: isPressed
+            isPressed: baseButtonState == .pressed
         )
     }
 
@@ -99,65 +98,51 @@ public struct VCheckBox<Label>: View where Label: View {
                 
             case .title(let title):
                 labeledCheckBoxView(label: {
-                    SwiftUIGestureBaseButton(
-                        onStateChange: stateChangeHandler,
-                        label: {
-                            Text(title)
-                                .multilineTextAlignment(uiModel.titleTextLineType.textAlignment ?? .leading)
-                                .lineLimit(type: uiModel.titleTextLineType.textLineLimitType)
-                                .minimumScaleFactor(uiModel.titleTextMinimumScaleFactor)
-                                .foregroundStyle(uiModel.titleTextColors.value(for: internalState))
-                                .font(uiModel.titleTextFont)
-                        }
-                    )
-                    .disabled(!uiModel.labelIsClickable) // `disabled(:_)` because it's a `SwiftUIGestureBaseButton`
+                    baseButtonView(label: { internalState in
+                        Text(title)
+                            .multilineTextAlignment(uiModel.titleTextLineType.textAlignment ?? .leading)
+                            .lineLimit(type: uiModel.titleTextLineType.textLineLimitType)
+                            .minimumScaleFactor(uiModel.titleTextMinimumScaleFactor)
+                            .foregroundStyle(uiModel.titleTextColors.value(for: internalState))
+                            .font(uiModel.titleTextFont)
+                    })
+                    .blocksHitTesting(!uiModel.labelIsClickable)
                 })
 
             case .label(let label):
                 labeledCheckBoxView(label: {
-                    SwiftUIGestureBaseButton(
-                        onStateChange: stateChangeHandler,
-                        label: {
-                            label(internalState)
-                        }
-                    )
-                    .disabled(!uiModel.labelIsClickable) // `disabled(:_)` because it's a `SwiftUIGestureBaseButton`
+                    baseButtonView(label: label)
+                        .blocksHitTesting(!uiModel.labelIsClickable)
                 })
             }
-        })
-        .applyIf(uiModel.appliesStateChangeAnimation, transform: {
-            $0.animation(uiModel.stateChangeAnimation, value: internalState)
         })
     }
     
     private var checkBoxView: some View {
-        SwiftUIGestureBaseButton(
-            onStateChange: stateChangeHandler,
-            label: {
-                ZStack(content: {
-                    RoundedRectangle(cornerRadius: uiModel.cornerRadius)
-                        .foregroundStyle(uiModel.fillColors.value(for: internalState))
-                    
-                    if uiModel.borderWidth > 0 {
-                        RoundedRectangle(cornerRadius: uiModel.cornerRadius)
-                            .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: uiModel.borderWidth)
-                    }
+        baseButtonView(label: { internalState in
+            ZStack(content: {
+                RoundedRectangle(cornerRadius: uiModel.cornerRadius)
+                    .foregroundStyle(uiModel.fillColors.value(for: internalState))
 
-                    if let checkmarkIcon {
-                        checkmarkIcon
-                            .applyIf(uiModel.isCheckmarkIconResizable, transform: { $0.resizable() })
-                            .applyIfLet(uiModel.checkmarkIconContentMode, transform: { $0.aspectRatio(nil, contentMode: $1) })
-                            .applyIfLet(uiModel.checkmarkIconColors, transform: { $0.foregroundStyle($1.value(for: internalState)) })
-                            .applyIfLet(uiModel.checkmarkIconOpacities, transform: { $0.opacity($1.value(for: internalState)) })
-                            .font(uiModel.checkmarkIconFont)
-                            .frame(size: uiModel.checkmarkIconSize)
-                    }
-                })
-                .frame(dimension: uiModel.dimension)
-                .clipShape(.rect(cornerRadius: uiModel.cornerRadius)) // Prevents large content from overflowing
-                .padding(uiModel.checkboxHitBox)
-            }
-        )
+                if uiModel.borderWidth > 0 {
+                    RoundedRectangle(cornerRadius: uiModel.cornerRadius)
+                        .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: uiModel.borderWidth)
+                }
+
+                if let checkmarkIcon: Image = checkmarkIcon(internalState: internalState) {
+                    checkmarkIcon
+                        .applyIf(uiModel.isCheckmarkIconResizable, transform: { $0.resizable() })
+                        .applyIfLet(uiModel.checkmarkIconContentMode, transform: { $0.aspectRatio(nil, contentMode: $1) })
+                        .applyIfLet(uiModel.checkmarkIconColors, transform: { $0.foregroundStyle($1.value(for: internalState)) })
+                        .applyIfLet(uiModel.checkmarkIconOpacities, transform: { $0.opacity($1.value(for: internalState)) })
+                        .font(uiModel.checkmarkIconFont)
+                        .frame(size: uiModel.checkmarkIconSize)
+                }
+            })
+            .frame(dimension: uiModel.dimension)
+            .clipShape(.rect(cornerRadius: uiModel.cornerRadius)) // Prevents large content from overflowing
+            .padding(uiModel.checkboxHitBox)
+        })
     }
 
     private func labeledCheckBoxView<Content>(
@@ -165,28 +150,38 @@ public struct VCheckBox<Label>: View where Label: View {
     ) -> some View
         where Content: View
     {
-        HStack(spacing: 0, content: {
+        HStack(spacing: uiModel.checkBoxAndLabelSpacing, content: {
             checkBoxView
-            spacerView
             label()
         })
     }
 
-    private var spacerView: some View {
-        SwiftUIGestureBaseButton(
-            onStateChange: stateChangeHandler,
-            label: {
-                Rectangle()
-                    .fixedSize(horizontal: false, vertical: true)
-                    .frame(width: uiModel.checkBoxAndLabelSpacing)
-                    .foregroundStyle(.clear)
+    private func baseButtonView(
+        label: @escaping (VCheckBoxInternalState) -> some View
+    ) -> some View {
+        SwiftUIBaseButton(
+            uiModel: uiModel.baseButtonSubUIModel,
+            action: {
+                playHapticEffect()
+                state.setNextState()
+            },
+            label: { baseButtonState in
+                let internalState: VCheckBoxInternalState = internalState(baseButtonState)
+
+                label(internalState)
+                    .applyIf(uiModel.appliesStateChangeAnimation, transform: {
+                        $0
+                            .animation(uiModel.stateChangeAnimation, value: state)
+                            .animation(nil, value: baseButtonState == .pressed) // Pressed state isn't shared between children
+                    })
             }
         )
-        .disabled(!uiModel.labelIsClickable) // `disabled(:_)` because it's a `SwiftUIGestureBaseButton`
     }
 
     // MARK: Checkmark Icon
-    private var checkmarkIcon: Image? {
+    private func checkmarkIcon(
+        internalState: VCheckBoxInternalState
+    ) -> Image? {
         switch internalState {
         case .off, .pressedOff: nil
         case .on, .pressedOn: uiModel.checkmarkIconOn
@@ -194,17 +189,7 @@ public struct VCheckBox<Label>: View where Label: View {
         case .disabled: nil
         }
     }
-    
-    // MARK: Actions
-    private func stateChangeHandler(gestureState: GestureBaseButtonGestureState) {
-        isPressed = gestureState.didRecognizePress
-        
-        if gestureState.didRecognizeClick {
-            playHapticEffect()
-            state.setNextState()
-        }
-    }
-    
+
     // MARK: Haptics
     private func playHapticEffect() {
 #if os(iOS)