+++ Sources/VComponents/Components/State Pickers/Check Box/VCheckBoxState.swift
@@ -8,14 +8,14 @@
 import Foundation
 import VCore
 
-// MARK: - V Toggle State
+// MARK: - V Check Box State
 /// Enumeration that represents state, such as `off`, `on`, or `indeterminate`.
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)
 public typealias VCheckBoxState = GenericState_OffOnIndeterminate
 
-// MARK: - V Checkbox Internal State
+// MARK: - V Check Box Internal State
 /// Enumeration that represents state, such as `off`, `on`, `indeterminate`, `pressed`, or `disabled`.
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)