+++ Sources/VComponents/Components/State Pickers/Toggle Button (Stretched)/VStretchedToggleButton.swift
@@ -47,8 +47,8 @@ import VCore
 ///         )
 ///     }
 ///
-@available(tvOS, unavailable) // Doesn't follow HIG. No `SwiftUIGestureBaseButton`
-@available(watchOS, unavailable) // Doesn't follow HIG. No `SwiftUIGestureBaseButton`.
+@available(tvOS, unavailable) // Doesn't follow HIG
+@available(watchOS, unavailable) // Doesn't follow HIG
 @available(visionOS, unavailable) // Doesn't follow HIG
 public struct VStretchedToggleButton<Label>: View where Label: View {
     // MARK: Properties - UI Model
@@ -56,16 +56,14 @@ public struct VStretchedToggleButton<Label>: View where Label: View {
 
     // MARK: Properties - State
     @Environment(\.isEnabled) private var isEnabled: Bool
-    @State private var isPressed: Bool = false
-    @Binding private var state: VStretchedToggleButtonState
-    private var internalState: VStretchedToggleButtonInternalState {
+    @Binding private var state: VRectangularToggleButtonState
+    private func internalState(_ baseButtonState: SwiftUIBaseButtonState) -> VStretchedToggleButtonInternalState {
         .init(
             isEnabled: isEnabled,
             isOn: state == .on,
-            isPressed: isPressed
+            isPressed: baseButtonState == .pressed
         )
     }
-
     // MARK: Properties - Label
     private let label: VStretchedToggleButtonLabel<Label>
 
@@ -123,43 +121,53 @@ public struct VStretchedToggleButton<Label>: View where Label: View {
 
     // MARK: Body
     public var body: some View {
-        SwiftUIGestureBaseButton(
-            onStateChange: stateChangeHandler,
-            label: {
-                labelView
+        SwiftUIBaseButton(
+            uiModel: uiModel.baseButtonSubUIModel,
+            action: {
+                playHapticEffect()
+                state.setNextState()
+            },
+            label: { baseButtonState in
+                let internalState: VStretchedToggleButtonInternalState = internalState(baseButtonState)
+
+                labelView(internalState: internalState)
                     .contentShape(Rectangle()) // Registers gestures even when clear
                     .frame(height: uiModel.height)
                     .clipShape(.rect(cornerRadius: uiModel.cornerRadius)) // Prevents large content from overflowing
-                    .background(content: { backgroundView }) // Has own rounding
-                    .overlay(content: { borderView }) // Has own rounding
+                    .background(content: { backgroundView(internalState: internalState) }) // Has own rounding
+                    .overlay(content: { borderView(internalState: internalState) }) // Has own rounding
+                    .applyIf(uiModel.appliesStateChangeAnimation, transform: {
+                        $0
+                            .animation(uiModel.stateChangeAnimation, value: state)
+                            .animation(nil, value: baseButtonState == .pressed)
+                    })
             }
         )
-        .applyIf(uiModel.appliesStateChangeAnimation, transform: {
-            $0.animation(uiModel.stateChangeAnimation, value: internalState)
-        })
     }
 
-    private var labelView: some View {
+    private func labelView(
+        internalState: VStretchedToggleButtonInternalState
+    ) -> some View {
         Group(content: {
             switch label {
             case .title(let title):
-                titleLabelViewComponent(title: title)
+                titleLabelViewComponent(internalState: internalState, title: title)
 
             case .icon(let icon):
-                iconLabelViewComponent(icon: icon)
+                iconLabelViewComponent(internalState: internalState, icon: icon)
 
             case .titleAndIcon(let title, let icon):
                 switch uiModel.titleTextAndIconPlacement {
                 case .titleAndIcon:
                     HStack(spacing: uiModel.titleTextAndIconSpacing, content: {
-                        titleLabelViewComponent(title: title)
-                        iconLabelViewComponent(icon: icon)
+                        titleLabelViewComponent(internalState: internalState, title: title)
+                        iconLabelViewComponent(internalState: internalState, icon: icon)
                     })
 
                 case .iconAndTitle:
                     HStack(spacing: uiModel.titleTextAndIconSpacing, content: {
-                        iconLabelViewComponent(icon: icon)
-                        titleLabelViewComponent(title: title)
+                        iconLabelViewComponent(internalState: internalState, icon: icon)
+                        titleLabelViewComponent(internalState: internalState, title: title)
                     })
                 }
 
@@ -173,6 +181,7 @@ public struct VStretchedToggleButton<Label>: View where Label: View {
     }
 
     private func titleLabelViewComponent(
+        internalState: VStretchedToggleButtonInternalState,
         title: String
     ) -> some View {
         Text(title)
@@ -184,6 +193,7 @@ public struct VStretchedToggleButton<Label>: View where Label: View {
     }
 
     private func iconLabelViewComponent(
+        internalState: VStretchedToggleButtonInternalState,
         icon: Image
     ) -> some View {
         icon
@@ -195,7 +205,9 @@ public struct VStretchedToggleButton<Label>: View where Label: View {
             .frame(size: uiModel.iconSize)
     }
 
-    private var backgroundView: some View {
+    private func backgroundView(
+        internalState: VStretchedToggleButtonInternalState
+    ) -> some View {
         RoundedRectangle(cornerRadius: uiModel.cornerRadius)
             .scaleEffect(internalState.isPressedOffPressedOn ? uiModel.backgroundPressedScale : 1)
             .foregroundStyle(uiModel.backgroundColors.value(for: internalState))
@@ -207,21 +219,13 @@ public struct VStretchedToggleButton<Label>: View where Label: View {
     }
 
     @ViewBuilder 
-    private var borderView: some View {
+    private func borderView(
+        internalState: VStretchedToggleButtonInternalState
+    ) -> some View {
         if uiModel.borderWidth > 0 {
             RoundedRectangle(cornerRadius: uiModel.cornerRadius)
                 .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: uiModel.borderWidth)
                 .scaleEffect(internalState.isPressedOffPressedOn ? uiModel.backgroundPressedScale : 1)
-        }
-    }
-
-    // MARK: Actions
-    private func stateChangeHandler(gestureState: GestureBaseButtonGestureState) {
-        isPressed = gestureState.didRecognizePress
-
-        if gestureState.didRecognizeClick {
-            playHapticEffect()
-            state.setNextState()
         }
     }