+++ Sources/VComponents/Components/State Pickers/Toggle Button (Stretched)/VStretchedToggleButtonUIModel.swift
@@ -15,6 +15,16 @@ import VCore
 @available(visionOS, unavailable)
 public struct VStretchedToggleButtonUIModel {
     // MARK: Properties - Global
+    var baseButtonSubUIModel: SwiftUIBaseButtonUIModel {
+        var uiModel: SwiftUIBaseButtonUIModel = .init()
+
+        // This flag is what makes the component possible.
+        // Animation can be handled within the component,
+        // at a cost of sacrificing tap animation.
+        uiModel.animatesStateChange = false
+
+        return uiModel
+    }
     /// Height.
     /// Set to `48` on `iOS`.
     /// Set to `40` on `macOS`.